backend/pyproject.toml
[tool.poetry]
name = "ftb-backend"
version = "0.1.0"
description = ""
authors = ["kst164 <kartheek@tammana.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
